Version for experiments shown in the thesis
