--- src/__init__.py
@@ -1,3 +1,0 @@
-
-
-from .code_search_manager import CodeSearchManager
